Planning this out.
On token retrieval, also get a list of spreadsheets that match
a particular pattern and save that list to a local db (along with
a timestamp).

When done writing spreadsheet titles to the database, update
the ListPrerence for the spreadsheet chooser to display a
"Last Updated: YYYYMMDD-HHSS" message in the subtitle/summary.

On chosing a spreadsheet, save the contents to the local db as
use that as seed for the remainder of the application--Quickplay
and such.